pyproject.toml
[tool.poetry]
name = "django-backend"
version = "0.1.0"
description = ""
authors = ["Artem Voskresensky <metravs@gmail.com>"]
readme = "README.md"
packages = [{include = "django_backend"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.6"
djangorestframework = "^3.14.0"
django-environ = "^0.9.0"
cpython = "^0.0.6"
psycopg2-binary = "^2.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
